(#27) Changed unit to double and display with .2 precision

+++ wio/JsonSerializable.h
@@ -7,7 +7,7 @@
 class JsonSerializable
 {
     public:
-        virtual std::string ToJson(float value, std::string measurementTime) {
+        virtual std::string ToJson(double value, std::string measurementTime) {
             return "{\"value\": " 
                 + std::to_string(value) 
                 + ", \"dateTime\": \""
